question-45-flask-1.0.4-p-2.7-in-docker-master/requirements.txt
Click==7.0
Flask==1.0.4
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
Flask-SQLAlchemy==2.4.3
flask-api==2.0
